requirements.txt
python-dotenv==1.1.0
requests==2.32.3
yfinance==0.2.61
modelcontextprotocol==0.1.0
